requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
git-filter-repo==2.47.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==6.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.42
twilio==9.7.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.20.1
Babel
